Add coloring example...sort of.
For whatever reason, if I try to render both walls and backgrounds,
the walls don't come out.  Tried to fix it a bit, but I wonder if maybe
ChunkyPNG has changed versions since the code in the book was written?
Whatever the cause, I've added padding around the background rectangles,
so there are small dots in each cell.  It sort of gets the point across.